Fix wasteful reruns of application beforeModel
I noticed that the feature detection search was running
whenever I navigated to a search result, it turns out
that refreshModel on the region query parameter meant
the application route was reloading on calls to
service:router#transitionTo because the calls didn’t
include the region query parameter. Rather than needing
to add the region to every transitionTo call, this
just lets the route load again but skips the various
detection/setup queries.