Restore tests, excise problem template sections
The tests no longer hang when these sections of the
AllocationRow component are removed… what does it mean?

This was my next step after I failed to find anythting of use
with getSettledState… it showed pending waiters and timers,
but peeking into the timer didn’t reveal what it was other
than probably being a Ember Concurrency task.