tools: structs method generation
Developers are required to implement methods like `Copy`, `Equals`, `Diff`,
and `Merge` for many of the types in the `structs` package, and this is both
time-consuming and error prone, having resulted in several correctness bugs
from failing to copy objects retrieved from go-memdb or failing to compare
objects during plans.

This changeset introduces a prototype tool to use with `go:generate`
directives that can generate methods automatically for our developers. Future
changesets will include documentation for developers and porting the existing
methods to use this tool. We'll continue to debug and refine the tool as we
go.

Co-authored-by: Derek Strickland <derek.strickland@gmail.com>
Co-authored-by: Tim Gross <tgross@hashicorp.com>