add assertion to rulesForName to prevent using namespaces without bound property
Namespaces changed from being a global property to page-level. As a result, when we use ember-can
to check for permissions to do operations like scale-job we want to check if the namespace the
job is under has permissions to perform that action if we're using a client token. Its very easy to
forget to pass a bound property to ember-can and we can have a way to throw an error to prevent a
bug from being committed. Our best option is using an assert.