refactor reservedAmount to be calculated based on Allocation model
Originally, we calculated RreservedAmount in the PrimaryMetric::Allocation component to be based
on the tracker service, however, this component is directly tied to the Allocation model and
gets the model passed in as a named argument in both places that its invoked, the Allocation Details page
and the Topology view. It should be safe to make this change and this does not break any existing tests.