network: modify CreateNetwork impls for func signature change.
The in-tree implementations of DriverNetworkManager.CreateNetwork
have been modified to account for the function signature change.
In updating for the change, the enhancement of adding hostnames to
network namespaces has also been added to the Docker driver, whilst
the default Linux manager does not current impelement it.

The allocrunner includes additional validation of the new parameter
to ensure it is used in the correct situations.