wip lifecycle restart alloc
Add lifecycle hook and sidecar to task runner config so we can easily
check if a task is a lifecycle task (and if so, which one:
pre/poststart, poststop) or a sidecar task. Also swallows the error when
trying to restart a one-shot lifecycle task (but adds trace message) and
then also adds the task name to the error message of the you can't
restart a stopped task error to make it more helpful. This part seems a
little janky since I just turned a var into a func but it works?