edited dependent keys in job client status helper
The job client status helper relies on updating whenever an allocation
client status updates. Or if a new allocation is added to the client
node. We were previously observing the allocations array as a whole,
but we're depending on a specific property in that list to update in
order to make sure we're getting notifications to update at the right
time.