feat: Add always_fetch flag to artifact stanza
This commit introduces a new flag to `artifact` stanza: `always_fetch`.
When an allocation restarts, the artifacts are not redownloaded if the
allocation directory already has the artifacts available. In some
scenarios, it is however desired to download the artifact file whenever
an allocation restarts, so setting this flag to `true` should make that
happen.