edit static figures, denominator for stats
The task resource utilization chart was breaking when cores was selected, as well.
This was a challenging situation because we mutate the AllocatedResources property
on Allocations in the Ember Data Serializer to merge all Tasks nested within
Allocated Resources for one combined Allocated Resource stat, to work around this,
we are getting the denominators from TaskStates which is a property that we're
computing and adding in information from the AllocatedResource before merging all
the tasks.