record hanging tests - current hunch... this has to deal with fragmentArrays
My best guess here is that because, we're using fragmentArrays. These
records do not have have ids. In the cases where our tests are hanging
we're using a computed property with @each to observe each element in
the array and checking for a specific proprety to change. Likely because
we're filtering. https://github.com/rwjblue/ember-data.model-fragments#filtered-record-arrays